Add gradient-free dead-head distribution script + fix bf16 attribution
dead_heads_distribution.py: computes the per-layer distribution of dead
attention heads for a model+suite and renders one bar chart per task in the
*_dead_heads.png style. Uses a forward-only per-head activation-magnitude
criterion (||z_head|| < 0.15 x layer-max), so it works on a CPU-offloaded
70B where a full backward pass is impractical. Also dumps all per-layer dead
counts and per-head norms to JSON so the raw data is preserved (not just PNGs).

active_subgraph_dot.py: fix compute_per_head_scores dtype bug — the layer-0
anchor was cast to float32 and then flowed through bfloat16 weights, raising
"expected mat1 and mat2 to have the same dtype" and zeroing all head scores
on bf16 models (e.g. Llama-70B). Keep the anchor in native dtype.

+++ active_subgraph_dot.py
@@ -661,7 +661,11 @@ def compute_per_head_scores(
     fwd_hooks = []
 
     def _anchor(act, hook):
-        return act.detach().float().requires_grad_(True)
+        # Keep the anchor in the model's native dtype.  Casting to float32 here
+        # injects a float32 tensor into bfloat16 matmuls on bf16 models (e.g.
+        # Llama-70B loaded in bfloat16), raising "expected mat1 and mat2 to have
+        # the same dtype".  Gradients are cast to float at scoring time below.
+        return act.detach().requires_grad_(True)
     fwd_hooks.append(("blocks.0.hook_resid_pre", _anchor))
 
     for l in range(n_layers):
